Make tests limiter implementation agnostic 1/2
The ovarall goal is to be able to run the same tests on all possible
implementations so that we can verify they're working.

After #42 we can now configure limiters with the same options.

After this diff, the tests only pass in those options in to the testRunner.

A following diff will modify testRunner so that all tests are executed
on both atomic & mutex based limiters.

This in turn allows us to change the limiters in sync & unblocks
resolving #27 & #23 (they are about the same bug)

This diff also
- changes the test to run in the same package.
  This is just so that I can access `option` without making it public,
  which seems unnecessary.
- renames runner into testRunner since it appears the benchmarks
  already define a `runner`. We might be able to merge them
  together but that's for another diff.