Fix irregularity in benchmark
The benchmark exetutes the test from 1 to 65k goroutines. Not to take
forever it appropriately uses steps as the number of goroutines grows.

The bucked between 1024 and 2048 has an irregularly low step.
I suspect this is because it's where the atomic limiter becomes
much slower.
However, the small step is not really necessary to show the degredation,
it'll be clearly visible anyway.

Let's make the steps bigger, and make the benchmark run faster.