Run benchmarks on push to master (#51)
This is so that we have history of the performance before
we start making changes.

Ideally, we'd have some kind of human readable display/numbers,
but that's another step.

Opting for running this only on push to master because those tests
are heave & take minutes - I don't think there's a need to run
them on every push. Might be wrong, but lets see how this goes.