Run tests on all (both) implementations
The test runner will simply execute all tests on all implementations.

This is so that we can:
- make sure both implementations are equivalent
- start adding more implementations
- have proper test coverage, yay.

The diff is slightly bigger than the actual changes:
- I'm modifying the tests to run in the same package as the code,
  so that I have access to each of the implementations.
- I'm renaming `runner` to `testRunner` since benchmark defines it's
  own runner.

I considered somehow keeping the `ratelimit_test` package,
and also having a separate `ratelimit` tests to just test
implementations  but that feels too complicated for the benefit - lets
keep it simple and just have a single set of tests.

We could probably also make the benchmark tests re-use the same test runner,
but that's a separate follow up.