Allow configuring custom slack
Resolves #20

This is to allow people to configure custom slack - we previously
defaulted to 10, and only allowed to disable it.

The tests are a bit painful to read, I had a stab at generating
the same set of tests for both with, and without per, but that was
even more hard to read.