Fix slack ignoring "Per" option
This is a split off #64 to make #64 smaller/easier.

I believe out current treatment of the default (10) slack doesn't
work very well with Per option.

We configure the limiter with an limit:
- `Limiter(10)` - means, allow 10 requests per second, with 10 requests
  slack - "slack period" is 1 second

However, if we do:
- `Limiter(10, Per(time.Minute))` we get - 10 requests per minute,
  but the slack is still 10 *per second*, meaning the slack period is
  stil 10 second - effectively 1 request.
I believe this is confusing, because with "slack of 10", I'd expect
a 1 minute "slack period".

After this change:
- `Limiter(10, Per(time.Minute))` we get - 10 requests per minute,
  and a slack of 10 requests - so, slack period is 1 minute.

We're basically making it so that both "rate" and "slack" are expressed
in the same unit (ints) and that "Per" applies to both of them.

Note this is currently of minor impact since we're hardcoding the
slack as 10. This will become way more painful/complicated with #64.