Update go directive in go.mod files, run `go mod tidy`
With #86 I thought bumping the directive would reduce the nose in the
go.sum files. It helps a bit, but less than I expected.
Perhaps it's worth trying anyway - we enable some module prunning for
packages that import us, I believe (?).

My understanding after talking to @abhinav is that this is fully
backwards compatible.